add dep and start project structure refactoring
