+++ version/version.go
@@ -1,0 +1,5 @@
+package version
+
+var (
+	Version = "0.0.1"
+)
